Revert "Obtain updated version from master"

+++ src/main/java/com/furnitureapp/repository/sales/SaleRepository.java
@@ -5,6 +5,6 @@ import com.furnitureapp.repository.IRepository;
 
 import java.util.Set;
 
-public interface SaleRepository extends IRepository<Sale, Integer> {
+public interface SaleRepository extends IRepository<Sale, Long> {
     Set<Sale> list();
 }
